client/requirements.txt
numpy>=1.24.0
requests>=2.31.0
websockets>=12.0
pyyaml>=6.0
sounddevice>=0.4.6
